refactor(web): extract useTimelines hook from App.jsx (sc-1651)
Sixth and final Phase-A slice — the most coupled. Moves timeline state (timelines,
timelinesProjectId, selectedTimelineId, activeTimeline), the two timeline refs
(selectedTimelineIdRef, timelineApplyQueueRef), all 10 timeline functions, and the
two timeline-specific effects (selectedTimelineId ref-sync, then load-on-selection,
in that order) into hooks/useTimelines.js. App keeps the SSE job.updated handler and
calls the returned enqueueTimelineGenerationApply; the bulk reset/project-load
effects use the returned setters/refreshTimelines. createVideoJob (App-owned) is
injected for the timeline generate-clip action; ensureItemVersionFields moved with
applyTimelineGenerationResult. EditorScreen prop signatures unchanged.

Behavior-preserving: App.jsx 1820 -> 1590 lines (god module 2334 -> 1590, -744 over
six slices). 107 web vitest pass, build clean, browser smoke (Rust API on a temp
data dir): created a timeline (createTimeline -> activeTimeline renders), and after a
reload the project-load refreshTimelines -> ref-sync -> load effect chain auto-loaded
the timeline (select shows it, editor renders it) with no console errors.
